lint max number of rows
